Do not serialize JDK classes with zero-config Compact serialization
It was decided to fail with meaningful exception messages in case
we are faced with a field type that is not supported by the
reflective serializer yet.

To do that, the team decided to exclude all JDK clases from the
zero-config serialization, meaning, they cannot be used as types,
field types, or array component types in the reflective serializers.