Introduce two phased schema replication mechanism
The former implementation was relying on eventual replication
of the schemas to all cluster members but it was susceptible to
total schema (hence data) loss under certain scenarios.

Also, it was duplicating schemas on each HotRestart/WAN entry
through a relatively expensive process.

The new implementation separates the replication process into two
phases.

- Preparation
- Acknowledgment

where in the preparation phase we make sure that the schema is available
in the in-memory registry(and HotRestart data, if available) of each
cluster member before moving on to the acknowledgment phase.

With the help of clients/members who make sure that schema is acknowledged
in the cluster before putting any data associated with it, this process
makes sure that schemas are available in each cluster member
before any data associated with it is put into the cluster.

The implementation has mechanisms to
- correctly replay replications occurred to the joining members
- heal after split-brain so that both splits of the cluster can
merge their schemas
- persist to HotRestart per schema (instead of per entry)
- replicate schemas per batch for WAN (instead of per event), with
the async acknowledgment mechanism to not send known to be replicated
schemas in the target cluster

The details of the implementation are available in greater detail
in the TDD

https://hazelcast.atlassian.net/wiki/spaces/IMDG/pages/3976593439/Compact+Schema+Distribution+TDD