Improve documentation for primitive-nullable primitive field interoperability
We allow reading nullable primitive fields as primitive fields
as long as they are not null. Also, it is possible to read
primitive fields as nullable primitives. The same holds true
for the arrays of such types.

This PR improves the Javadoc we have for these methods and mentions
about this possibility.