Improve handling of SQLNonTransientException in WriteJdbcP (#21931)
In `WriteJdbcP#process` we check for SQLNonTransientException, but there
can be a chain of exceptions via getNextException(), where a
SQLNonTransientException can be present.

Also in `WriteJdbcP#connectAndPrepareStatement` we call
`javax.sql.DataSource#getConnection` which can throw
SQLNonTransientException, in such case we just keep retrying.

This PR fixes and unifies the check for SQLNonTransientException.
It improves the state, but it really depends on JDBC driver implementing
the SQLNonTransientException correctly. E.g. PostgreSQL throws only
PSQLException and doestn't carry any information if the failure is
transient (apart from the server error).