Add support for List, Set, and Map fields to zero-config serializers
This PR adds support for List (and ArrayList), Set (and HashSet),
and Map (and HashMap) fields to the zero config serializers.

List and Set are written as the arrays of the component types they have
(like List<String> -> ARRAY_OF_STRING).

Maps are written as two lists back to back, with the key and value
type of the Map. The field names of these two lists are fieldName + Keys
and fieldName + Values respectively.

Also, this PR changes the reflective and record serializer so
that they re-use the same code.

address review comments

add more fields with interface collection types