Check nested fields of the Compact serializable objects are not serializable by other mechanisms in zero-config serialization
It could be the case that the top level class is serializable by
Compact, but it might have nested fields, which can also be serialized
by other mechnasims like Java serialization.

The problem is, we cache the serializers, even the reflective
serializers. And, if we cache that we can serialize Java serializable
objects with Compact, this can be problematic for such scenarios

```
map.put(javaSerializable, 1)
map.put(zeroConfigSerializableWithJavaSerializableField, 2)
map.get(javaSerializable) // this returns null now, as the
// key was serialized with identified, but now it is Compact
```

To solve this, we check if the nested field is only be
serializable with Compact (either there is no other serilization
mechanism, or Compact serializer is explicitly serialized with
it with `addClass` or `addSerializer`).