Defer post-join op execution during recovery
While recovery from persistence is in progress, post-join ops should
be deferred, to be executed after recovery is done. This is enforced
for all operations in OperationRunnerImpl#checkNodeState, however post-
join ops are executed directly, within OnJoinOp which is allwoed during
passive state, therefore is not caught by conditions in checkNodeState.