Increase test timeout in the CompactSchemaReplicationStressTest
The test was failing regularly due to timeout errors. In the failing
runs, there were no kind of exceptions etc, and the test logic is
pretty simple. So, the test is probably failing due to timeout.

On my laptop, the test takes around 2 minutes to complete, and in
some nightly builds that I have examined, it was taking around
4 minutes. So, I think it is reasonable to think that test could
take more than 5 minutes in some nightly builds.

Therefore, in this PR, I increase the test timeout to 10 minutes,
and added logging to the drivers so that we can see the progress
in the logs, to make sure that the failures were indeed related
to timeouts.