Improve the client connection process listener (#21833)
We have merged the code for the client connection process listener
for MC, but there were some improvements possible.

Formerly,
- the code had an unnecessary level of abstractions. It was hard
to understand what is `ListenerAggregate`, or which listener
wraps which one etc.
- some listener methods were executed in the connection manager's
executor, but some were not.
- the exception handling and logging around the implementation
had some mistakes

To fix it, I have introduced a runner class, that will execute
all listeners in a single-threaded executor (which is created
only if a listener is added, so no executor is created by default).
The error handling and possible address translations are moved
to that class. We now have a list of listeners on this one, and
no confusing wrapping occurs anymore.