Wait for submitted tasks during the shutdown in the ClientConnectionProcessListenerRunner
The connection diagnostics related events are fired in a separate
executor.

We shutdown that executor during the regular client shutdown. However,
we were using `executor.shutdownNow`, which clears the task queue
of the executor, so nothing more than the currently executing task
is executed during the shutdown.

However, this was wrong. It is quite possible that the client fires
a terminal connection diagnostics event like `clusterConnectionFailed`,
after which it quickly calls shutdown. Now, it might be possible to
not fire the listener for this event, because we put it to the
queue of the executor, but called shutdownNow, quickly after it.

To fix it, we now use `executor.shutdown` in the stop method of
the listener runner, which will wait for the completion of the
already submitted tasks, with the help of `executor.awaitTermination`.

Also, did some auto-formatting changes and put the listener invocation
in a try-catch to possibly avoid throwing `RejectedExecutionException`
for the events offered after the client shutdown. (I don't believe it
is the case today, but added it nevertheless to be safe)