Handle default package in the zero-config Compact serializers better for Java 8
In Java 9 and above, for the classes in the default package, the
`clazz.getPackage()` method returns a non-null package with an
empty class name.

However, it returns a null package in Java 8, for the same classes.

We were saying that a class is not-zero-config-serializable if the
package was null (because it is null if it is a primitive type, array
or void in Java 9+, which are not supported in the paths that we do
the checks in this method). That was not correct in Java 8.

Now, if the package is null, we are performing the checks done in
Java 9+ and above again to distinguish classes in the default package.