Send client state to cluster in re-connections
In split-brain scenarios, the cluster id stays the same for
different halves of the split.

When the client disconnects from the first half it was connected
to and reconnects to the other half, from the point of view of
the client, it connects to the same cluster.

However, it might very well be the case that, the client has sent
some state to the first half of the cluster, and that state must
be replicated to the other half when the client reconnects to it.

This is especially needed for Compact schemas and user code
deployment classes.

In this PR, we are sending the client state to the cluster after
reconnections, regardless it is connected back to possibly the same
cluster with the same id or not.