Wait for disconnection in the ClientReconnectTest#testCallbackAfterServerShutdown
In this test, a putAsync call is made after the server is shut down.

However, if the putAsync request is made before the connection on
the client-side is dropped, we would get a TargetDisconnectedException,
instead of the expected HazelcastClientNotActiveException.

To fix it, we should make sure that the server connection is dropped
on the client-side, before making the putAsync request.