Fix possible overtaking of barriers by items [HZ-1394] (#22207)
The scenario was this. In one `drainTo` call we observed a barrier on all queues
but one. In the next call we observed the remaining barrier, but there was also
an item in a queue after this barrier. When the barrier was observed, we cleared
the `receivedBarriers` bitset, but because we continued to the next queue, if
there was a normal item, it was returned before the barrier waiting in the
`specialItemStash`.

The failing test passed 200 times, while it failed every 20th-40th run before
the fix locally for me.

Fixes #21908