Don't allow overriding default serializers with Compact
Overriding default serializers with Compact might allow easier use
for some of the field types, but it can lead to quite subtle
problems to debug, when used incorrectly.

Therefore, we have decided to not allow it for 5.2, and this
PR removes the integration of the Compact serialization with the
`allowOverrideDefaultSerializers` configuration option.