Invoke schema replication operations in parallel
We were using `InvocationUtil#invokeOnStableClusterSerial` to
invoke schema replication related operations in all cluster members
in serial, restarting the iterations in case of topology changes.

It works fine, but for the requirements of the schema replication,
the process can happen in parallel as well, making it more scalable
for large clusters.

This PR introduces a counterpart of the aforementioned method to
run the given operation in stable a cluster in parallel, exhibiting
the same behavior in case of exceptional cases.

The only difference of the new method is that it skips executing
the operation on this member, to make the implementation in the
schema replicator cleaner. If in the future, someone needs
a variant of this method that needs to run the operation in
this member as well, we can simply add a boolean flag for it.

Also, this PR replaces all usages of CompletableFuture in the
Compact related parts with the InternalCompletableFuture
to make sure that methods of the form `xxx` and
`xxxAsync(without the executor)` run in the default executor.
In the CompletableFuture, the place to run the callbacks was
undefined, so it is better to be safe and use what we use
in the rest of the code base.