Add check that server and hz-cli versions match [HZ-1173] (#22224)
Jet compatibility guarantees require the hz-cli and server versions to
be the same minor version. This commit adds check for all hz-cli
operations.

It might be possible that we need to ignore the check (e.g. for testing
or debugging), for such case a --ignore-version-mismatch option is
provided.

Fixes #20019

Breaking changes (list specific methods/types/messages):

- previously it was possible to submit a job using a different version of hz-cli from the server version. Generally, it would work when there were no change in serialized form of the pipeline. There is no guarantee. Now we fail early.