Support splittable for CSV format in file connector [HZ-1405] (#22221)
When the input file is large it is split into multiple FileSplits.
Previously the CsvInputFormat wasn't respecting the split boundaries and
read the whole file. Causing duplicates when the file had multiple 
splits.

Migrated to LineRecordReader, which handles splits and is suitable to
read file line by line, required for CSV.

Additional logic was added to the `initialize` method to handle the first
line of the CSV file containing the header. It is used either to remap
the order of fields or to build a schema.

Also added tests for non-splittable (gzip) and splittable (bzip2)
compression.

Fixes #21854