Fix avoiding the buffer in stream-to-stream join
We incorrectly handled the case when we received an item, but it's not possible
that it will join any future row from the other side. We correctly avoided
adding it to the buffer, but we didn't check if it could possibly join something
already in the buffer, so some join output was lost. The fix is to avoid the
buffer, but still try to join the already-buffered rows.

I also refactored how the minimum buffer time is calculated - this wasn't
strictly needed, but in the process I thought that I will need the minimum
buffer time for the evaluation of the above. It's more efficient because we
already full-scan the buffer when clearing it, and when removing from the buffer
it's the only place where the minimum can increase. We decrease the minimum when
adding to the buffer.