Fix cooperative emission in stream-to-stream join (#22335)
When multiple rows in the buffer matched, and we couldn't emit them all because
the outbox was full, we'll add the last item to `pendingOutput`. The next call
to `tryProcess` will emit the `pendingOutput`, but then it will return `true`
and miss the remaining matches in the buffer, causing incorrect results.