[HZ-1208] merge persistence & hot-restart (MapConfig)
Additionally, when looking up config via
lookupByPattern, configs must be cloned and named
after the looked-for name, similarly to ConfigUtils#getConfig.
Otherwise, returned config has the name of the template
config (eg "default" or "persistent.*") and this results
in strange failures (like HZ-1208) with persistence.

With these fixes, it is revealed that assertions
in ConfigSearchTest and DynamicConfigSearchOrderTest are faulty, since
they are based on config name. The tests are now fixed to compare
on other properties of the config (eg backupCount).