Use BIG_ENDIAN byte order for the serializer id in toDataWithSchema
In Hazelcast, the top level serializer id is always encoded in
BIG_ENDIAN byte order, and for nested fields, the configured
byte order is used.

toDataWithSchema was not following this behaviour, and was using
the configured byte order for the top level class.

Note that, for nested compacts, we don't recursively call toDataWithSchema.
The logic to serialize nested fields are handled within the serializer.