Add destroy permission to DestroyRaftObjectMessageTask
Formerly, the message task was returning null for the required
permissions, signaling that destroying the CP objects requires
no permission.

However, it is possible to configure DESTROY permission for such
objects as well. And, it is definitely an unexpected thing to
be able to destroy a CP object without this permission, when
the security and permission config is enabled for such objects.

This PR adds the correct permissions to the required message task.

The tests will be added to the enterprise PR.