Fix SARG nullAs issues (#22418)
The index part is backported from
https://github.com/hazelcast/hazelcast/pull/22238.

The not-index part:
The ```SearchableExpression``` is serialized with IDS. It is a part of
```DAG``` or ```ExecutionPlan```. If we want to add any byte into
serialized form of the object in a way that the previous version can
deserialize it, the new piece of data must be added at the end of byte
stream.

The introduced workaround tracks all created instances of
```SearchableExpression``` and adds additional data to serialized form
of the top-level object (```DAG``` or ```ExecutionPlan```). The
serialized form of such object can be described as:
```
[DAG serialized bytes from 5.1.4][new bytes]
```
During the deserialization on 5.1.[0-4] the IDS will take only ```[DAG
serialized bytes from 5.1.4]``` and deserialize it properly. If a new
version is deserializing old format of the class the ```EOFException```
will be thrown. We handle that exception as a natural situation and that
is the way to know that the data is coming from the previous version.

Some thoughts about the solution:
- it is nasty, I don't like it
- the solution in master is much better, so we need to maintain that
hack in ```5.1.z``` line only
- the issue is severe IMO so I believe we should merge it

Checklist:
- [x] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [x] Label `Add to Release Notes` or `Not Release Notes content` set
- [x] Request reviewers if possible

Co-authored-by: Viliam Durina <viliam@hazelcast.com>