Improve Jackson dependency declaration (#22617)
Jackson sometimes releases micro-patch for a CVE, e.g. 2.13.4.2.

This version is not available for all Jackson modules, only for the
affected module. The bom with this version is published in format
x.y.z.YYYYMMDD.

We import Jackson bom in root pom.xml, but we use ${jackson.version}
property in some modules. This means we can't use single version
property when we need to use such bom.

This PR changes remove all uses of ${jackson.version} property apart
from the bom import. All Jackson dependency declarations are without
version, using the version imported by bom in the root pom.xml.

Users declaring a dependency on a module also correctly get the version
from the bom, e.g. project declaring a dependency on
`hazelcast-jet-cdc-debezium`, which has a dependency on
jackson-annotations has the following in its dependencies:

```
[INFO]    \- com.fasterxml.jackson.core:jackson-annotations:jar:2.13.4:compile
```

The version is taken from the bom.

Addresses CVEs from #22407 in direct and transitive dependencies, where
Jackson isn't shaded by the 3rd party.