Fix member leave test - wait for job deployment (#22476)
The test worked like that:

1. execute sql query on ```instance1```
2. wait for receiving any row
3. shutdown ```instance2``` in separate thread
4. expect query to fail with ```MemberLeftException```

The fact that we received a row doesn't mean that the job is already
deployed on ```instance2```. If we shutdown that instance before that
member starts the job we get ```HazelcastInstanceNotActiveException```,
not the ```MemberLeftException```.

I changed the second step, we do not wait for any row, but we wait for
job to be executing on both members before performing shutdown.

Fixes https://github.com/hazelcast/hazelcast/issues/20501

Checklist:
- [x] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [x] Label `Add to Release Notes` or `Not Release Notes content` set
- [x] Request reviewers if possible

(cherry picked from commit 7320c147922cdee18056282847f7a41b0910a1be)