Convert HotRestartTriggerBackupMessageTask and ShutdownClusterMessageTask to async message tasks (#22737)
Formerly, the tasks were callable, meaning that they will send a
response
back to the client, once the call method completes.

There was a fundamental problem with these tasks' implementation.
They were scheduling a task to be executed in the execution service,
and depending on the result of it, sending a response or an exception
to the client.

However, since the tasks were callable, we were actually sending a
response to the client, once the call method finishes.

After the execution service completes the runnable, it was sending
another response or exception, depending on the result of the
call, but the client was ignoring it, because it was receiving
a successful response for that correlation id before.

That's why we were not able to notify clients about the
actual status of the operations. The second response, whether
it is an exception or normal response, was ignored by the client.

As a solution, I converted these tasks to async tasks, which
handle the proper error handling themselves and send the
correct response type according to the actual result of the
runnable tasks.