Fix GenericMapStore showing UndefinedErrorCodeException on the client-side [HZ-1737]
In some cases, if there is a server-side exception when using a map w/ a GenericMapStore, on the client-side the exception is observed as com.hazelcast.client.UndefinedErrorCodeException

Expected behavior
A defined exception should be thrown such as com.hazelcast.sql.HazelcastSqlException, which can be handled accordingly by the user.

It is happening because the client side is trying to recreate the exceptions with a constructor having (String message, Throwable cause)  signature. If the constructor does not exist, the client side is showing UndefinedErrorCodeException.
In order to show HazelcastSqlException and QueryException, two new constructors are added

closes : https://github.com/hazelcast/hazelcast/issues/22740