Fix @Category annotations on tests [5.2.z]
Using @Category on a subclass doesn't replace the categories on
superclass, but it adds to the categories. So if you declare QuickTest
and ParallelJVMTest on parent and only QuickTest on child it will run in
parallel phase in PR builder. Fixed all cases like this in hazelcast
module.

If the child uses either SlowTest or NightlyTest it is ok, because it
will run during respective executions and Quick/Parallel are not
considered at all, so I left these cases as they are.

Possibly affected tests: #21318, #21156

Backport of #22837