Fix @Category annotations on tests (#22837)
Using `@Category` on a subclass doesn't replace the categories on
superclass, but it adds to the categories. So if you declare QuickTest
and ParallelJVMTest on the parent and only QuickTest on the child it
will run in parallel phase in PR builder. Fixed all cases like this in
`hazelcast` module.

If the child uses either SlowTest or NightlyTest it is ok, because it
will run during respective executions and Quick/Parallel are not
considered at all, so I left these cases as they are.

This has been highlighted when we updated surefire/failsafe to 3.0.0-M7,
previously used version 2.22.2 has a bug and would ignore these tests.

Possibly affected tests: #21318, #21156