Fix com.hazelcast.jet.kafka.impl.StreamKafkaPTest.when_eventsInAllPartitions_then_watermarkOutputImmediately [HZ-1773]
In the test logs there was an error message as below.
"Unable to get partition metadata, ignoring: org.apache.kafka.common.errors.TimeoutException: Timeout expired while fetching topic metadata"
This indicates that Kafka was not ready when the processor was trying to read metadata. The test was publishing to Kafka asynchronously. The fix is changing this and publishing to Kafka synchronously, assuming that if the call succeeds, Kafka partitions should be ready. Then processor should be able to access metadata

closes : https://github.com/hazelcast/hazelcast/issues/22858