Use client drivers in ClientQueryCacheBounceTest [API-1661] (#22867)
The test was not setting the driver to be used, and was using everything
from the superclass as it is.

It was causing the test to use member drivers. To solve it, I made the
driver type configurable and set it to the client driver type only in
the ClientQueryCacheBounceTest.