Make ClientXATest use mock network API-1657 (#22901)
The test failed with a log that includes explicit suspicion requests.
Instance1 that is used to start transaction received explicit suspicion
requests from the other members and it closed connections to those
members. The cluster size did not become 3 as it supposed to be. Also
there are logs like these in the logs:

`New join request has been received from current master address. The
UUID in the join request (62a29dcc-f128-4b24-b442-5f6a09e64dfe) is
different from the known master one (08032ec1-...`

and there is no member with uuid 62a29dcc-f128-4b24-b442-5f6a09e64dfe
among the 3 members.

I think the problem is the same as with the one discussed at
https://hazelcast.slack.com/archives/C01JU7ZJYGP/p1668592206511029

Note that this test uses real network but I don't think the runner is
wrongly set for this test. It runs in serial, not parallel. I wonder if
we can port this test to mock network. Any idea why it uses a real
network?

The fix converts the test to use HazelcastFactory and testRecovery waits
for cluster size to be 3 before starting to run actual code.
Will be backported to all other branches until 4.0.x, because the test
is really old.

Fixes #22271