Use assertTrueEventually in AuthenticationInformationLeakTest API-1652 (#22857)
The test failure was due to the client connection being closed before
the exception message was able to be written. Upon closing, write queues
of the connection being drained and this means no exception message will
be sent to the client. To fix this, I used assertTrueEventually in the
test.

Alternative ways of fixing this test failure is that:

1. We can delay closing the connection in AbstractMessageTask, (we think
this is not the ideal solution)
2. We can change the logic so that we dont close the client connection
at all.

Will be backported to 5.2.z as well upon merging. 

Fixes #22564