Fix throttling interval calculation for SQL [HZ-1413] (#22641)
Fixes #21983 : assign the GCD of all the window sizes in the query plan, or 100ms otherwise. Also, S2S join is an exception: therefore, we use a hard-coded value that's a trade-off between latency and the number of watermarks. The default maximum throttling size is 100, but we reduce it to one-tenth of the minimum join bounds spread if that's less.

Co-authored-by: Burak Gök <burakgok95@gmail.com>
Co-authored-by: Viliam Durina <viliam@hazelcast.com>